requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
google-generativeai
langchain-huggingface
faiss-cpu
python-dotenv
sentence-transformers
